guilt: remove GIT_EXEC_BASE export
git-core provides the following routines (among others): git-sh-setup,
cd_to_toplevel, die, and more.  But it is not always in the same location
relative to the guilt binary if git is not part of a 'host tools'
sysroot. Modern git versions don't need this, so commenting it out (until it
breaks again) is the solution to the problem of it not existing, and the
library routines will sourced and provided by git itself.

Since bitbake.conf has git-native as ASSUME_PROVIDED, this means that
when the system git binary is used, errors such as this can be seen in
failed patch logs:

| /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/guilt: line 29: /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/../libexec/git-core/git-sh-setup: No such file or directory
| /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/guilt: line 717: cd_to_toplevel: command not found
| error: patch failed: Makefile:2
| error: Makefile: patch does not apply
| /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/guilt-push: line 137: die: command not found
| [ERROR] unable to complete push
| pending patches are:
| /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/guilt: line 29: /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/../libexec/git-core/git-sh-setup: No such file or directory
| /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/guilt: line 29: /opt/poky/build/tmp/sysroots/x86_64-linux/usr/bin/../libexec/git-core/git-sh-setup: No such file or directory
| links/linux-yocto-custom/0001-linux-version-tweak.patch
| ERROR. could not update git tree

With the export removed, we have a normal "clean" failure message when
the sysroot or system git is used.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>