distcc: updated to version 3.1
Added two new flags to the configure script:

--disable-Werror: don't treat all warnings all errors (which breaks compilation).

PYTHON=/dev/null: this prevents distcc from detecting the host Python and trying
to build its include server using the host Python. This disables the include
server completely. If the include server is needed, that should be the object of
another patch (and would introduce a dependency on python for distcc).

The 'distcc-avahi' and 'makefile-param-order' patches are not needed anymore, as
they were merged upstream.

Signed-off-by: Bogdan Marinescu <bogdan.a.marinescu@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>