mulitilib_header: Avoid sstate checksum issues for -native recipes
The use of arch specific variables like MIPSPKGSFX_ABI was creeping
into the -native sstate checksums of package like ncurses-native.

This is pointless and undesireable. We could add specific variable
exclusions but we might as well just brute force the code to be disabled
in the -native case since we don't use multilibs in the native case.

[YOCTO #3827]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>