v86d: compile against userspace safe kernel headers
v86d was building directly against the staged kernel in the target
sysroot. This has been fine in the past, but with recent (3.7) changes
to user headers into the "uapi" include structure, there are no longer

Without the preprocesor protection, v86d fails to build against any
kernel with uapi header files.

v86d doesn't actually need the whole kernel tree, the exported headers
are enough to build a working binary. This change sets the v86d variable
to ${STAGING_DIR_HOST}/usr, and once the v86d build adds "include" to the
end of KDIR, we have a valid include path to the target sysroot headers.

This also works for pre-uapi kernel headers, so the change is safe for
most (if not all) kernel's with properly exported headers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>