util-linux: Add package for libmount
The libmount shared library was part of util-linux. This caused
util-linux-mount to RDEPEND on util-linux, so including that would
also drag in all of util-linux and all its recommendations.

To break this circular dependency, add a libmount package that holds
the libmount.so.* files, just like the other libraries built by util-
linux.

Signed-off-by: MiLo <milo-software@users.sourceforge.net>
Signed-off-by: Saul Wold <sgw@linux.intel.com>