man: Remove wildcard from SRC_URI
[ CQID: WIND00402802 ]

Remove the wildcard from the SRC_URI.  This causes problems when you .bbappend
and add a FILESEXTRAPATHS entry.  The unpack task may be unable to find the
files to unpack leading to an error.

Avoid wildcards at all costs...

Note, remove both man-1.5m2-multiple.patch and man-1.6e-lzma+xz-support.patch,
they were not being applied in the previous version.  The later also fails to
apply when added to the SRC_URI.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>