recipes: Fix ALLOW_EMPTY with no package specified
There are various usages of ALLOW_EMPTY with no packages specified. This
is not recommended syntax, nor is it likely to be supported in the future.
This patch improves the references in OE-Core, either removing them if they're
pointless (e.g. when PACKAGES="") or specifying which package it applies to.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>