distro: Add separatebuilddir.inc
Going forward its going to be useful to separate build data from source data
in those autotooled projects which support it. Unfortunately there is a lot
of breakage so for now, this starts the creation of an opt in list which
we can iterate over enable more recipes over time.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>