perl: avoid splitting out .debug directories as packages
The recursive parameter is set to True, the regex for the first call to
do_split_packages matches any path under ${libdir}/perl/${PV}/auto/, and
the .debug directories contain .so files, so each one was getting picked
up as a package. Change the regex to disallow dots in the path beneath
auto/ and thus avoid the .debug directories.

Fixes [YOCTO #4048].

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>