rpm: Fix debugedit buildid processing
[ YOCTO #4089 ]

When constructing a new buildid, the items being hashed need to be
returned to their native endian.  In the process we were munging
the sh_type field that we relied on to determine if a section was
loadable or not.  The patch avoids this behavior by only modifying
a copy of the local endian data.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>