linux-yocto/3.8: update mips SA_RESTORER fix
During the 3.8.4 integration there was a build issue on MIPS due to
SA_RESTORER changes. A solution was put in place for mips, but it
didn't cover other impacted architectures.

This is a backport of the proposed fix for the next 3.8-stable,
since the full -stable might not be available in the right timeframe.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>